add cat & add sub cat ok

+++ src/Controller/AdCategoryController.php
@@ -17,9 +17,13 @@ class AdCategoryController extends AbstractController
 {
     /**
      * @Route("/", name="ad_category_index", methods={"GET"})
+     * @param AdCategoryRepository $adCategoryRepository
+     * @return Response
      */
     public function index(AdCategoryRepository $adCategoryRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('ad_category/index.html.twig', [
             'ad_categories' => $adCategoryRepository->findAll(),
         ]);
@@ -27,9 +31,14 @@ class AdCategoryController extends AbstractController
 
     /**
      * @Route("/new", name="ad_category_new", methods={"GET","POST"})
+     * @param Request $request
+     * @return Response
      */
     public function new(Request $request): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $adCategory = new AdCategory();
         $form = $this->createForm(AdCategoryType::class, $adCategory);
         $form->handleRequest($request);
@@ -38,6 +47,8 @@ class AdCategoryController extends AbstractController
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($adCategory);
             $entityManager->flush();
+
+            $this->addFlash('success', 'Successfully Added');
 
             return $this->redirectToRoute('ad_category_index');
         }
@@ -50,9 +61,14 @@ class AdCategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="ad_category_show", methods={"GET"})
+     * @param AdCategory $adCategory
+     * @return Response
      */
     public function show(AdCategory $adCategory): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('ad_category/show.html.twig', [
             'ad_category' => $adCategory,
         ]);
@@ -60,14 +76,22 @@ class AdCategoryController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="ad_category_edit", methods={"GET","POST"})
+     * @param Request $request
+     * @param AdCategory $adCategory
+     * @return Response
      */
     public function edit(Request $request, AdCategory $adCategory): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(AdCategoryType::class, $adCategory);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
+
+            $this->addFlash('success', 'Successfully Updated');
 
             return $this->redirectToRoute('ad_category_index');
         }
@@ -80,13 +104,19 @@ class AdCategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="ad_category_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param AdCategory $adCategory
+     * @return Response
      */
     public function delete(Request $request, AdCategory $adCategory): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$adCategory->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($adCategory);
             $entityManager->flush();
+            $this->addFlash('success', 'Successfully Deleted');
         }
 
         return $this->redirectToRoute('ad_category_index');
