+++ src/Controller/AdSubCategoryController.php
@@ -3,7 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\AdSubCategory;
-use App\Form\AdSubCategory3Type;
+use App\Form\AdSubCategoryType;
 use App\Repository\AdSubCategoryRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -17,9 +17,13 @@ class AdSubCategoryController extends AbstractController
 {
     /**
      * @Route("/", name="ad_sub_category_index", methods={"GET"})
+     * @param AdSubCategoryRepository $adSubCategoryRepository
+     * @return Response
      */
     public function index(AdSubCategoryRepository $adSubCategoryRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('ad_sub_category/index.html.twig', [
             'ad_sub_categories' => $adSubCategoryRepository->findAll(),
         ]);
@@ -27,17 +31,23 @@ class AdSubCategoryController extends AbstractController
 
     /**
      * @Route("/new", name="ad_sub_category_new", methods={"GET","POST"})
+     * @param Request $request
+     * @return Response
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $adSubCategory = new AdSubCategory();
-        $form = $this->createForm(AdSubCategory3Type::class, $adSubCategory);
+        $form = $this->createForm(AdSubCategoryType::class, $adSubCategory);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($adSubCategory);
             $entityManager->flush();
+
+            $this->addFlash('success', 'Successfully Added');
 
             return $this->redirectToRoute('ad_sub_category_index');
         }
@@ -50,9 +60,13 @@ class AdSubCategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="ad_sub_category_show", methods={"GET"})
+     * @param AdSubCategory $adSubCategory
+     * @return Response
      */
     public function show(AdSubCategory $adSubCategory): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('ad_sub_category/show.html.twig', [
             'ad_sub_category' => $adSubCategory,
         ]);
@@ -60,14 +74,21 @@ class AdSubCategoryController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="ad_sub_category_edit", methods={"GET","POST"})
+     * @param Request $request
+     * @param AdSubCategory $adSubCategory
+     * @return Response
      */
     public function edit(Request $request, AdSubCategory $adSubCategory): Response
     {
-        $form = $this->createForm(AdSubCategory3Type::class, $adSubCategory);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        $form = $this->createForm(AdSubCategoryType::class, $adSubCategory);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
+
+            $this->addFlash('success', 'Successfully Updated');
 
             return $this->redirectToRoute('ad_sub_category_index');
         }
@@ -80,13 +101,19 @@ class AdSubCategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="ad_sub_category_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param AdSubCategory $adSubCategory
+     * @return Response
      */
     public function delete(Request $request, AdSubCategory $adSubCategory): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$adSubCategory->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($adSubCategory);
             $entityManager->flush();
+            $this->addFlash('success', 'Successfully Deleted');
         }
 
         return $this->redirectToRoute('ad_sub_category_index');
